# swagger.yaml
openapi: 3.0.0
info:
  title: Express API with Swagger
  version: 1.0.0
paths:
...
  /chat/member/{channelId}:
    get:
      tags:
        - Chat
      summary: Get channel members by channel ID
      description: Get channel members by channel ID
      parameters:
        - in: path
          name: channelId
          required: true
          description: ID of the chat channel
          schema:
            type: string
      responses:
        '200':
          description: Channel members retrieved successfully
  /chat/last-message/{channelId}:
    get:
      tags:
        - Chat
      summary: Get last message in a channel by channel ID
      description: Get last message in a channel by channel ID
      parameters:
        - in: path
          name: channelId
          required: true
          description: ID of the chat channel
          schema:
            type: string
      responses:
        '200':
          description: Last message retrieved successfully
  /chat/messages/{channelId}:
    get:
      tags:
        - Chat
      summary: Get messages in a channel by channel ID
      description: Get messages in a channel by channel ID
      parameters:
        - in: path
          name: channelId
          required: true
          description: ID of the chat channel
          schema:
            type: string
        - in: query
          name: skip
          required: false
          description: Number of messages to skip
          schema:
            type: integer
        - in: query
          name: limit
          required: false
          description: Maximum number of messages to retrieve
          schema:
            type: integer
        - in: header
          name: Authorization
          required: true
          description: Bearer token for authentication
          schema:
            type: string
      responses:
        '200':
          description: Messages retrieved successfully
  /chat/channel/find:
    post:
      tags:
        - Chat
      summary: Find or create a chat channel
      description: Find or create a chat channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                memberIds:
                  type: array
                  items:
                    type: string
              example:
                memberIds:
                  - userId1
                  - userId2
      responses:
        '200':
          description: Channel found or created successfully
...